+++ src/BugsnagUnity/Clients/AndroidClient.cs
