+++ src/BugsnagUnity/Clients/Client.cs
@@ -6,117 +6,6 @@ using UnityEngine.SceneManagement;
 
 namespace BugsnagUnity
 {
-  public delegate void Middleware(Report report);
-
-  public interface IClient
-  {
-    IConfiguration Configuration { get; }
-
-    IBreadcrumbs Breadcrumbs { get; }
-
-    ISessionTracker SessionTracking { get; }
-
-    User User { get; }
-
-    void Send(IPayload payload);
-
-    Metadata Metadata { get; }
-
-    void BeforeNotify(Middleware middleware);
-
-    void Notify(System.Exception exception);
-
-    void Notify(System.Exception exception, Middleware callback);
-
-    void Notify(System.Exception exception, Severity severity);
-
-    void Notify(System.Exception exception, Severity severity, Middleware callback);
-  }
-
-  interface INativeClient
-  {
-    /// <summary>
-    /// The native configuration
-    /// </summary>
-    IConfiguration Configuration { get; }
-
-    /// <summary>
-    /// The native breadcrumbs
-    /// </summary>
-    IBreadcrumbs Breadcrumbs { get; }
-
-    /// <summary>
-    /// The native delivery method
-    /// </summary>
-    IDelivery Delivery { get; }
-
-    /// <summary>
-    /// Populates the native app information
-    /// </summary>
-    /// <returns></returns>
-    void PopulateApp(App app);
-
-    /// <summary>
-    /// Populates the native device information
-    /// </summary>
-    /// <returns></returns>
-    void PopulateDevice(Device device);
-
-    /// <summary>
-    /// Adds the metadata to the native clients metadata
-    /// </summary>
-    /// <param name="metadata"></param>
-    void SetMetadata(string tab, Dictionary<string, string> metadata);
-
-    /// <summary>
-    /// Populates the native user information.
-    /// </summary>
-    /// <returns></returns>
-    void PopulateUser(User user);
-
-    /// <summary>
-    /// Populates any native metadata.
-    /// </summary>
-    /// <param name="metadata"></param>
-    void PopulateMetadata(Metadata metadata);
-  }
-
-  class NativeClient : INativeClient
-  {
-    public IConfiguration Configuration { get; }
-
-    public IBreadcrumbs Breadcrumbs { get; }
-
-    public IDelivery Delivery { get; }
-
-    public NativeClient(Configuration configuration)
-    {
-      Configuration = configuration;
-      Breadcrumbs = new Breadcrumbs(configuration);
-      Delivery = new Delivery();
-    }
-
-    public void PopulateApp(App app)
-    {
-    }
-
-    public void PopulateDevice(Device device)
-    {
-    }
-
-    public void PopulateMetadata(Metadata metadata)
-    {
-    }
-
-    public void PopulateUser(User user)
-    {
-    }
-
-    public void SetMetadata(string tab, Dictionary<string, string> metadata)
-    {
-    }
-  }
-
   class Client : IClient
   {
     public IConfiguration Configuration => NativeClient.Configuration;
