+++ src/BugsnagUnity/Clients/MacOsClient.cs
