+++ src/BugsnagUnity/Clients/iOSClient.cs
